move handling functions to the parent component, add confirm popup Relates #52
