+++ client/src/Pages/DashboardProvider/Orders/index.js
@@ -1,17 +1,18 @@
 import React, { useState, useEffect } from 'react';
 
 import Axios from 'axios';
-import { Alert, Spin, Tabs } from 'antd';
+import { Alert, Spin, Tabs, Modal } from 'antd';
+import { ExclamationCircleOutlined } from '@ant-design/icons';
 
 import PendingProvider from './pending';
 import errorHandel from '../../../Utils/errorHandel';
 
 const { TabPane } = Tabs;
+const { confirm } = Modal;
 
 const Orders = () => {
   const [ordersData, setOrdersData] = useState([]);
   const [isLoading, setIsLoading] = useState(false);
-  const [refresh, setRefresh] = useState(false);
   const [error, setError] = useState('');
 
   useEffect(() => {
@@ -35,10 +36,40 @@ const Orders = () => {
     return () => {
       unmounted = false;
     };
-  }, [refresh]);
+  }, []);
 
-  const handleRefresh = () => {
-    setRefresh(!refresh);
+  const handleAcceptOrder = async (orderId) => {
+    try {
+      setIsLoading(true);
+      await Axios.post(`/api/v1/user/order-requests/${orderId}`, {
+        time: '',
+      });
+      setIsLoading(false);
+    } catch (err) {
+      errorHandel(setError, err);
+    }
+  };
+
+  const handleCancelOrder = async (orderId) => {
+    confirm({
+      title: 'Are you sure delete this order?',
+      icon: <ExclamationCircleOutlined />,
+      okText: 'Yes',
+      okType: 'danger',
+      cancelText: 'No',
+      async onOk() {
+        try {
+          await Axios.delete(`/api/v1/user/order-requests/${orderId}`);
+          setOrdersData(ordersData.filter(({ id }) => id !== orderId));
+        } catch (err) {
+          errorHandel(setError, err);
+        }
+      },
+    });
+  };
+
+  const handleMoreDetails = () => {
+    // open popup modal
   };
 
   return (
@@ -58,9 +89,8 @@ const Orders = () => {
                   mobile,
                   state,
                   description,
-                  id,
                   date,
-                  order_id: orderId,
+                  id,
                 }) => ({
                   userinfo: [username, avatar],
                   location,
@@ -69,10 +99,13 @@ const Orders = () => {
                   description,
                   key: id,
                   time: date,
-                  action: orderId,
+                  action: id,
                 })
               )}
-              refresh={handleRefresh}
+              handleCancelOrder={handleCancelOrder}
+              handleAcceptOrder={handleAcceptOrder}
+              handleMoreDetails={handleMoreDetails}
+              error={error}
             />
           </TabPane>
           <TabPane tab="Orders" key="2">
